pyproject.toml
[tool.poetry]
name = "website"
version = "1.0.0"
description = ""
authors = ["Jay Cho <jaycho1214@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
psycopg2-binary = "^2.9.9"
google-cloud-storage = "^2.14.0"
gunicorn = "^21.2.0"
sqlmodel = "^0.0.14"
sentry-sdk = {extras = ["django"], version = "^1.39.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
